use size_t in strncasecmp

+++ lib/common/utils.h
@@ -29,7 +29,7 @@ extern "C" {
     extern int strcasecmp(const char *s1, const char *s2);
 #endif
 #ifndef HAVE_STRNCASECMP
-    extern int strncasecmp(const char *s1, const char *s2, unsigned int n);
+    extern int strncasecmp(const char *s1, const char *s2, size_t n);
 #endif
 
     extern void *zmalloc(size_t);
